Improvements have been made to the search for strike cells and work has begun with the chessboard
The method for searching for possible hits has been improved. If earlier there was a “head-on” method, now it has been improved by creating vectors of the required size, which improves operating time.
Documentation describing the methods has also been added to the Figure.hpp file.
In the file Table_of_Chess, the creation of a chessboard and the placement of pieces on it has begun.