The final version of chess!
Fixed bugs with invalid moves. Updated the way to store chess coordinates from a vector to an unordered_set. The corresponding change was made for all functions that used vector. To improve the architecture, constexpr string has been changed to constexpr string_view.  Methods for checking the validity of a move and the necessary functions for this have also been added.